requirements.txt
streamlit
tensorflow
Pillow==9.4.0
numpy
python-multipart
